Update view loader flow

+++ packages/@pinjs/view/index.js
@@ -1,9 +1,10 @@
+const fs = require('fs');
 const path = require('path');
 const webpack = require('webpack');
 const build = require('./libs/build');
 const utils = require('./libs/utils');
 const logger = require('./libs/logger');
-const dev = require('./dev');
+const isProduction = process.env.NODE_ENV === 'production';
 
 class PinView {
     constructor(config) {
@@ -14,45 +15,54 @@ class PinView {
             serverOutputDir: path.join(process.cwd(), 'build/server'),
         }, config || {});
 
-        this.webpackConfig = {};
-        this.webpackDevServerOptions = {};
-        this.webpackCompiler = null;
-
         this.SSRBundleManifest = null;
         this.SSRBuildpath = this.config.serverOutputDir;
         this.SSRBuildClass = null;
         this.SSRBuild = null;
+
+        if (!isProduction) {
+            this.webpackConfig = {};
+            this.webpackOptions = {};
+            this.webpackCompiler = null;
+            this.webpackDevMiddleware = null;
+        }
     }
 
     async loadSSRBuild() {
-        this.SSRBundleManifest = require(this.config.serverOutputDir + '/react-loadable-manifest.json');
         this.SSRBuildFile = this.SSRBuildpath + '/main.js';
+        this.SSRBundleManifestFile = this.config.serverOutputDir + '/react-loadable-manifest.json';
+
+        if (!fs.existsSync(this.SSRBuildFile) || !fs.existsSync(this.SSRBundleManifestFile)) {
+            logger.error('No production build found!');
+            process.exit();
+        }
+
         utils.purgeCache(this.SSRBuildFile);
+        this.SSRBundleManifest = require(this.SSRBundleManifestFile);
         this.SSRBuildClass = require(this.SSRBuildFile);
         await this.SSRBuildClass.preload();
         this.SSRBuild = new this.SSRBuildClass(this.config);
     }
 
-    async init(rebuild = true, webpackDevServerOptions = {}) {
-        this.webpackDevServerOptions = dev.config(this.config.publicPath, webpackDevServerOptions);
+    async init(webpackOptions = {}) {
+        if (isProduction) {
+            return await this.loadSSRBuild();
+        }
+
+        this.webpackOptions = utils.config(this.config.publicPath, webpackOptions);
         this.webpackConfig = await build.getWebpackConfigs(this.config);
         this.webpackCompiler = webpack(this.webpackConfig);
+        this.webpackDevMiddleware = await utils.getWebpackDevMiddleware(this.webpackCompiler, this.webpackOptions.devServer);
 
-        if (rebuild) {
-            await build.buildClient(this.config, this.webpackCompiler.compilers[0]);
-            await build.buildServer(this.config, this.webpackCompiler.compilers[1]);
-            logger.info('> Build completed')
-        }
-        await this.loadSSRBuild();
-        await this.initWebpack();
-    }
-
-    async initWebpack() {
         const clientCompiler = this.webpackCompiler.compilers[0];
         const serverCompiler = this.webpackCompiler.compilers[1];
-        await dev.getHotClient(clientCompiler, this.webpackDevServerOptions.hotClient);
+        await utils.getWebpackHotClient(clientCompiler, this.webpackOptions.hotClient);
+        await this.loadSSRBuild();
+        this.initWebpack(serverCompiler);
+    }
 
-        serverCompiler.hooks.watchRun.tapAsync('pinjsView', (_compiler, done) => {
+    async initWebpack(serverCompiler) {
+        serverCompiler.hooks.watchRun.tapAsync('PinJsView', (_compiler, done) => {
             let watchFileSystem = _compiler.watchFileSystem;
             let watcher = watchFileSystem.watcher || watchFileSystem.wfs.watcher;
             let updatedFile = Object.keys(watcher.mtimes)[0];
@@ -62,14 +72,10 @@ class PinView {
             }
             return done();
         });
-        serverCompiler.hooks.done.tapAsync('pinjsView', async (stats, done) => {
+        serverCompiler.hooks.done.tapAsync('PinJsView', async (stats, done) => {
             await this.loadSSRBuild();
             return done();
         });
-    }
-
-    middleware() {
-        return dev.middleware(this.webpackCompiler, this.webpackDevServerOptions.middleware);
     }
 
     async render(req, res, pagePath, query) {
@@ -77,8 +83,19 @@ class PinView {
             return res.end('ok');
         }
 
-        const rendered = await this.SSRBuild.render(pagePath, {}, this.SSRBundleManifest);
-        res.end(`<!doctype html><html lang="en"><head>${rendered.cssFiles.map(file => `<link rel="stylesheet" href="${file}" />`)}</head><body><div id="app">${rendered.html}</div>${rendered.jsScripts.join('\n')}</body></html>`);
+        await utils.beforeViewRender(req, res, this.webpackDevMiddleware);
+
+        let rendered = await this.SSRBuild.render(pagePath, {}, this.SSRBundleManifest);
+        let cssScripts = rendered.cssScripts || [];
+        let jsScripts = rendered.jsScripts || [];
+
+        if (!isProduction && !this.webpackOptions.devServer.writeToDisk) {
+            let devAssets = utils.getSSRDevMiddlewareAssets(pagePath, res.locals.webpackStats, res.locals.fs);
+            cssScripts = devAssets.cssScripts || [];
+            jsScripts = devAssets.jsScripts || [];
+        }
+
+        res.end(`<!doctype html><html lang="en"><head>${cssScripts.join('\n')}</head><body><div id="app">${rendered.html}</div>${jsScripts.join('\n')}</body></html>`);
     }
 }
